Revert "Battleship api"
