+++ .gitignore
@@ -14,6 +14,3 @@ Thumbs.db
 
 # Logs
 npm-debug.log
-
-#build
-dist
