--- eslint.config.js
@@ -1,37 +1,0 @@
-import eslintPluginPrettier from 'eslint-plugin-prettier';
-import eslintConfigPrettier from 'eslint-config-prettier';
-import typescriptPlugin from '@typescript-eslint/eslint-plugin';
-import parser from '@typescript-eslint/parser';
-
-export default [
-  {
-    files: ['**/*.ts'],
-    languageOptions: {
-      parser,
-      parserOptions: {
-        ecmaVersion: 'latest',
-        sourceType: 'module',
-        project: './tsconfig.json',
-      },
-    },
-    plugins: {
-      '@typescript-eslint': typescriptPlugin,
-      prettier: eslintPluginPrettier,
-    },
-    rules: {
-      ...typescriptPlugin.configs.recommended.rules,
-      ...eslintConfigPrettier.rules,
-      'prettier/prettier': 'error',
-      '@typescript-eslint/no-unused-vars': [
-        'warn',
-        { argsIgnorePattern: '^_' },
-      ],
-      '@typescript-eslint/explicit-function-return-type': 'off',
-      '@typescript-eslint/no-explicit-any': 'warn',
-      'no-console': 'warn',
-    },
-  },
-  {
-    ignores: ['dist/**', 'node_modules/**'],
-  },
-];
