+++ package.json
@@ -1,37 +1,21 @@
 {
-  "name": "ws_task",
-  "version": "1.0.0",
-  "description": "Template for the WebSocket remote control task",
-  "main": "src/index.ts",
-  "scripts": {
-    "start": "webpack --mode=production && node dist/bundle.js",
-    "start:dev": "nodemon --watch \"src/**/*.ts\" --exec \"ts-node-dev --respawn --transpile-only\" src/index.ts",
-    "format": "prettier --write \"src/**/*.ts\"",
-    "lint": "eslint \"src/**/*.ts\" --fix",
-    "prepare": "npm run lint && npm run format",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "Andrei Auchynnikau",
-  "keywords": [],
-  "license": "ISC",
-  "devDependencies": {
-    "@eslint/js": "^9.26.0",
-    "@eslint/json": "^0.12.0",
-    "@eslint/markdown": "^6.4.0",
-    "@types/ws": "^8.18.1",
-    "eslint": "^9.26.0",
-    "eslint-config-prettier": "^10.1.5",
-    "eslint-plugin-prettier": "^5.4.0",
-    "nodemon": "^3.1.10",
-    "prettier": "^3.5.3",
-    "ts-loader": "^9.5.2",
-    "ts-node-dev": "^2.0.0",
-    "typescript": "^5.8.3",
-    "typescript-eslint": "^8.32.1",
-    "webpack": "^5.99.8",
-    "webpack-cli": "^6.0.1"
-  },
-  "dependencies": {
-    "ws": "^8.8.0"
-  }
+    "name": "ws_task",
+    "version": "1.0.0",
+    "description": "Template for the WebSocket remote control task",
+    "main": "index.js",
+    "type": "module",
+    "scripts": {
+        "start": "node ./index.js",
+        "start:dev": "nodemon ./index.js",
+        "test": "echo \"Error: no test specified\" && exit 1"
+    },
+    "author": "Andrei Auchynnikau",
+    "keywords": [],
+    "license": "ISC",
+    "devDependencies": {
+        "nodemon": "^2.0.16"
+    },
+    "dependencies": {
+        "ws": "^8.8.0"
+    }
 }
