--- src/errors/input-validation-error.ts
@@ -1,7 +1,0 @@
-export default class InputValidationError extends Error {
-  code: string;
-  constructor(message: string) {
-    super(message);
-    this.code = 'IVErr';
-  }
-}
