--- webpack.config.js
@@ -1,23 +1,0 @@
-const path = require("path")
-
-module.exports = {
-  entry: "./src/index.ts",
-  target: "node",
-  mode: "production",
-  output: {
-    filename: "bundle.js",
-    path: path.resolve(__dirname, "dist"),
-  },
-  resolve: {
-    extensions: [".ts", ".js"],
-  },
-  module: {
-    rules: [
-      {
-        test: /\.ts$/,
-        use: "ts-loader",
-        exclude: [/node_modules/, /src\/tests/]
-      },
-    ],
-  },
-}
